refactor: extract SettingsStore+DisplayMask extensions (#636 slice B)
Move 4 masked-display accessors + shared `private mask` helper from
SettingsStore (main class body) to a new SettingsStore+DisplayMask.swift
extension file. Byte-preserving; same pattern as merged slice A (#697/#853).

**Moved (54 source lines):**
- `maskedAPIKey` (8 lines)
- `maskedSelectedAIProviderCredential` (9 lines)
- `maskedGitHubToken` (8 lines)
- `maskedJiraAPIToken` (8 lines)
- `mask(secret:persistenceState:emptyPlaceholder:lockedPlaceholder:)` private helper (21 lines)

**Line-count contract met:**
- `SettingsStore.swift`: 1899 → 1845 (−54 lines)
- `SettingsStore+DisplayMask.swift`: 64 lines

**Visibility unchanged:** `mask(...)` stays `private` inside the extension file.

**Tests added:** `SettingsStoreDisplayMaskTests.swift` with 8 tests covering
2 persistence states (`.empty`, `.pendingSave`) × 4 accessors.

Closes #854
