+++ Sources/BugNarrator/Services/SettingsStore.swift
@@ -388,14 +388,6 @@ final class SettingsStore: ObservableObject {
         return assignedElsewhere ? nil : suggestion
     }
 
-    var maskedAPIKey: String {
-        mask(
-            secret: trimmedAPIKey,
-            persistenceState: apiKeyPersistenceState,
-            emptyPlaceholder: "No key saved",
-            lockedPlaceholder: "Saved key locked"
-        )
-    }
 
     var selectedAIProviderCredentialPersistenceState: APIKeyPersistenceState {
         guard aiProvider != .parakeetLocal else {
@@ -430,15 +422,6 @@ final class SettingsStore: ObservableObject {
         )
     }
 
-    var maskedSelectedAIProviderCredential: String {
-        let persistenceState = selectedAIProviderCredentialPersistenceState
-        return mask(
-            secret: persistenceState == .empty ? "" : trimmedAPIKey,
-            persistenceState: persistenceState,
-            emptyPlaceholder: aiProvider == .parakeetLocal ? "No key required" : "No key saved",
-            lockedPlaceholder: "Saved key locked"
-        )
-    }
 
 
 
@@ -732,14 +715,6 @@ final class SettingsStore: ObservableObject {
         )
     }
 
-    var maskedGitHubToken: String {
-        mask(
-            secret: trimmedGitHubToken,
-            persistenceState: githubTokenPersistenceState,
-            emptyPlaceholder: "No token saved",
-            lockedPlaceholder: "Saved token locked"
-        )
-    }
 
 
     var normalizedGitHubRepositoryOwner: String {
@@ -788,14 +763,6 @@ final class SettingsStore: ObservableObject {
         !trimmedJiraAPIToken.isEmpty
     }
 
-    var maskedJiraAPIToken: String {
-        mask(
-            secret: trimmedJiraAPIToken,
-            persistenceState: jiraTokenPersistenceState,
-            emptyPlaceholder: "No token saved",
-            lockedPlaceholder: "Saved token locked"
-        )
-    }
 
 
     var normalizedJiraBaseURL: String {
@@ -1601,27 +1568,6 @@ final class SettingsStore: ObservableObject {
     }
 
 
-    private func mask(
-        secret: String,
-        persistenceState: APIKeyPersistenceState,
-        emptyPlaceholder: String,
-        lockedPlaceholder: String
-    ) -> String {
-        guard !secret.isEmpty else {
-            switch persistenceState {
-            case .keychain:
-                return "Saved key"
-            case .keychainLocked:
-                return lockedPlaceholder
-            default:
-                return emptyPlaceholder
-            }
-        }
-
-        let suffixCount = min(4, secret.count)
-        let suffix = secret.suffix(suffixCount)
-        return "••••••••\(suffix)"
-    }
 
     private func prepareSecretsForUserInitiatedAccess(
         slots: [SecretSlot],
